+++ EasyEyes/EasyEyesActions.cs
@@ -10,7 +10,6 @@ public class EasyEyesActions : IEasyEyesActions
     private readonly Action _showOverlay;
     private readonly Action _hideOverlay;
     private readonly Action _showToast;
-    private readonly Action _clearToast;
 
     private TimeSpan _tRemaining;
     private readonly TimeSpan _tDuration;
@@ -25,7 +24,6 @@ public class EasyEyesActions : IEasyEyesActions
         Action showOverlay,
         Action hideOverlay,
         Action showToast,
-        Action clearToast,
         Action<Trigger> fireTrigger)
     {
         _timeProvider = timeProvider;
@@ -35,7 +33,6 @@ public class EasyEyesActions : IEasyEyesActions
         _showOverlay = showOverlay;
         _hideOverlay = hideOverlay;
         _showToast = showToast;
-        _clearToast = clearToast;
 
         _tTimer = new DispatcherTimer();
         _tTimer.Tick += (_, _) =>
@@ -56,7 +53,6 @@ public class EasyEyesActions : IEasyEyesActions
     public void ShowOverlay() => _showOverlay();
     public void HideOverlay() => _hideOverlay();
     public void ShowToast() => _showToast();
-    public void ClearToast() => _clearToast();
 
     public void SuspendTTimer()
     {