+++ EasyEyes/EasyEyesStateMachine.cs
@@ -39,7 +39,7 @@ public interface IEasyEyesActions
     void ShowOverlay();
     void HideOverlay();
     void ShowToast();
-    void ClearToast();
+
     void SuspendTTimer();
     void ResumeTTimer();
     void ResetTTimer();
@@ -145,13 +145,11 @@ public class EasyEyesStateMachine
             {
                 _actions.ResumeTTimer();
                 _actions.StopLTimer();
-                _actions.ClearToast();
             })
             .OnEntryFrom(Trigger.ScreenWake, () =>
             {
                 _actions.ResumeTTimer();
                 _actions.StopLTimer();
-                _actions.ClearToast();
             })
             .OnEntryFrom(Trigger.Resume, () =>
             {